+++ CHANGELOG.md
@@ -26,6 +26,7 @@ ChangeLog
 - BUGFIX: fix "xio-webrtc" node by catching errors on listening
 - BUGFIX: fix "xio-device" node by shutdown the stream on destruction
 - BUGFIX: fix "a2t-deepgram" chunk kind classification logic
+- BUGFIX: fix StreamWrapper utility class to correctly flush data
 - UPDATE: upgrade NPM dependencies and patches
 - CLEANUP: add missing type definitions for Kitten-TTS
 - CLEANUP: update year in all copyright messages